all components upgraded
